+++ playground/test.play.ts
@@ -1,0 +1,7 @@
+
+function hello(...args) {
+  console.log(args)
+}
+
+const aaa = []
+hello(...aaa)